% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrapping function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of bootstrap samples}

\item{x}{a vector of data}

\item{fun}{the function to be used to calculate the statistic}

\item{alpha}{the confidence level}

\item{cx}{the position of the confidence interval}

\item{...}{additional arguments to be passed to the hist function}
}
\value{
a vector that contains the confidence interval, the function used to calculate the statistic, and the data
}
\description{
Bootstrapping function
}
\examples{
myboot2(10000,1:10,"mean",0.05,1.5)
myboot2(10000,1:10,"median",0.05,1.5)
myboot2(10000,1:10,"sd",0.05,1.5)
}
