% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{Confidence interval for the mean}
\usage{
myci(x)
}
\arguments{
\item{x}{a vector of quantitative data}
}
\value{
a vector of length 2 with the lower and upper bounds of the confidence interval
}
\description{
Confidence interval for the mean
}
\examples{
myci(x=1:10)
}
