% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{Makes iter samples of n random numbers from a uniform distribution
and makes a histogram of the means of the samples.}
\usage{
myclt(n, iter, a = 0, b = 5)
}
\arguments{
\item{n}{the number of elements in each sample}

\item{iter}{the number of samples to take}

\item{a}{the lower bound of the uniform distribution}

\item{b}{the upper bound of the uniform distribution}
}
\value{
a vector of the means of the samples
}
\description{
Makes iter samples of n random numbers from a uniform distribution
and makes a histogram of the means of the samples.
}
\examples{
myclt(10, 1000)
myclt(10, 1000, 0, 10)
myclt(30, 1000, 0, 10)
}
