% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Max likelihood function}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{a likelihood function}

\item{x}{a vector of data}

\item{param}{a vector of parameter values}

\item{...}{additional arguments to be passed to plot}
}
\value{
a list with the index of the maximum likelihood, the parameter value at the maximum likelihood, the maximum likelihood, and the slope of the likelihood function at the maximum likelihood
}
\description{
For repeated sampling from same distribution
}
\examples{
mymaxlik(
       x=c(3,3,4,3,4,5,5,4),
       param=seq(0,1,length=1000),
       lfun=function(x,param,size) log(dbinom(x,prob=param,size=20)),
       xlab=expression(pi), main="Binomial",
       cex.main=2
       )
}
