% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.r
\name{myncurve}
\alias{myncurve}
\title{Plotting a normal distribution curve and finding the area under the curve for a given range. Also prints the area on the graph and prints the mean, std, and area as a list to the command line.}
\usage{
myncurve(mu = 0, sigma = 1, a = 0)
}
\arguments{
\item{mu}{mean of the normal distribution}

\item{sigma}{standard deviation of the normal distribution}

\item{a}{upper limit of the range}
}
\description{
Plotting a normal distribution curve and finding the area under the curve for a given range. Also prints the area on the graph and prints the mean, std, and area as a list to the command line.
}
\examples{
myncurve(0, 1, 2)
myncurve(0, 1, 1)
myncurve(5, 2, 3)
}
