% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myntickets.R
\name{myntickets}
\alias{myntickets}
\title{This function calculates the number of tickets to sell for a flight}
\usage{
myntickets(N = 400, gamma = 0.02, p = 0.95)
}
\arguments{
\item{N}{number of seats on the flight}

\item{gamma}{probability of the plane being truly overbooked}

\item{p}{probability of a passenger showing up}
}
\description{
This function calculates the number of tickets to sell for a flight
}
\examples{
myntickets(400, 0.02, 0.95)
myntickets(200, 0.02, 0.95)
myntickets(400, 0.01, 0.90)
}
