% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypvalue.R
\name{mypvalue}
\alias{mypvalue}
\title{Calculate the p-value for a t-test}
\usage{
mypvalue(t0, x, xmax = 4, n = 20, alpha = 0.05)
}
\arguments{
\item{t0}{the t value to test}

\item{x}{the vector of quantiles to plot}

\item{xmax}{the maximum x value to plot}

\item{n}{the degrees of freedom}

\item{alpha}{the alpha level for the test}
}
\value{
a list with the quantiles and the p-value
}
\description{
Calculate the p-value for a t-test
}
\examples{
mypvalue(2.5)
mypvalue(2.5, xmax=5)
mypvalue(2.5, xmax=5, n=30)
mypvalue(2.5, xmax=5, n=30, alpha=0.01)
}
