% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myquadratic.R
\name{myquadratic}
\alias{myquadratic}
\title{Get y value from a quadratic equation}
\usage{
myquadratic(a, b, c, x)
}
\arguments{
\item{a}{the coefficient of x^2}

\item{b}{the coefficient of x}

\item{c}{the constant}

\item{x}{the x value}
}
\value{
the y value of the quadratic equation at the given x value
}
\description{
Get y value from a quadratic equation
}
\examples{
myquadratic(1, 2, 3, 4)
}
