% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysample.R
\name{mysample}
\alias{mysample}
\title{Creates plots of sample distributions using replacement}
\usage{
mysample(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{the number of samples}

\item{iter}{the number of iterations}

\item{time}{the time to wait between iterations}
}
\description{
Creates plots of sample distributions using replacement
}
\examples{
mysample(n=1000, iter=30)
mysample(n=1000, iter=30, time=1)
}
